move detect_sources() to top level
